fix: fix wrapping and overflow in account groups list
