+++ package.json
@@ -1,115 +1,115 @@
 {
-    "type": "module",
-    "dependencies": {
-        "@azure/msal-browser": "^2.33.0",
-        "@azure/msal-react": "^1.5.3",
-        "@hookform/resolvers": "^3.9.0",
-        "@radix-ui/react-accordion": "^1.2.0",
-        "@radix-ui/react-checkbox": "^1.1.1",
-        "@radix-ui/react-dialog": "^1.1.1",
-        "@radix-ui/react-dropdown-menu": "^2.1.1",
-        "@radix-ui/react-popover": "^1.1.1",
-        "@radix-ui/react-progress": "^1.1.0",
-        "@radix-ui/react-scroll-area": "^1.1.0",
-        "@radix-ui/react-select": "^2.1.1",
-        "@radix-ui/react-toggle": "^1.1.0",
-        "@radix-ui/react-toggle-group": "^1.1.0",
-        "@radix-ui/react-tooltip": "^1.1.2",
-        "@sentry/react": "^9.0.0",
-        "@tanstack/react-query": "^5.59.0",
-        "class-variance-authority": "^0.7.0",
-        "clsx": "^2.1.1",
-        "commander": "^12.1.0",
-        "d3": "^7.9.0",
-        "date-fns": "^3.6.0",
-        "downshift": "^7.6.2",
-        "express": "^4.21.0",
-        "fuse.js": "^7.1.0",
-        "history": "^4.10.1",
-        "lodash-es": "^4.17.21",
-        "marked": "^14.1.2",
-        "minimist": "^1.2.8",
-        "numbro": "^2.5.0",
-        "radix-ui": "^1.1.3",
-        "react": "^18.3.1",
-        "react-dom": "^18.3.1",
-        "react-dropzone": "^14.2.3",
-        "react-helmet-async": "^2.0.5",
-        "react-hook-form": "^7.53.0",
-        "react-window": "^1.8.10",
-        "semver": "^7.6.3",
-        "styled-components": "^6.1.13",
-        "superagent": "^8.0.9",
-        "tailwind-merge": "^2.5.2",
-        "tailwindcss-animate": "^1.0.7",
-        "tsx": "^4.19.1",
-        "winston": "^3.17.0",
-        "wouter": "^3.3.5",
-        "yup": "^0.32.11",
-        "zustand": "^4.5.5"
-    },
-    "devDependencies": {
-        "@faker-js/faker": "^9.6.0",
-        "@sentry/vite-plugin": "^3.2.1",
-        "@storybook/addon-actions": "^8.6.3",
-        "@storybook/addon-essentials": "^8.6.3",
-        "@storybook/addon-interactions": "^8.6.3",
-        "@storybook/addon-links": "^8.6.3",
-        "@storybook/preview-api": "^8.6.3",
-        "@storybook/react": "^8.6.3",
-        "@storybook/react-vite": "^8.6.3",
-        "@storybook/test": "^8.6.3",
-        "@tailwindcss/vite": "^4.0.8",
-        "@testing-library/jest-dom": "^5.16.5",
-        "@testing-library/react": "^16.0.1",
-        "@testing-library/user-event": "^14.5.2",
-        "@types/d3": "^7.4.3",
-        "@types/lodash-es": "^4.17.12",
-        "@types/react": "^18.3.5",
-        "@types/react-dom": "^18.3.0",
-        "@types/react-window": "^1.8.8",
-        "@types/styled-components": "^5.1.34",
-        "@types/superagent": "^8.1.9",
-        "@typescript-eslint/eslint-plugin": "^8.5.0",
-        "@typescript-eslint/parser": "^8.5.0",
-        "@vitejs/plugin-react": "^4.3.2",
-        "@vitest/eslint-plugin": "^1.1.32-beta.3",
-        "eslint": "^8.57.1",
-        "eslint-config-prettier": "^9.1.0",
-        "eslint-plugin-react": "^7.37.4",
-        "eslint-plugin-react-hooks": "^5.1.0",
-        "jsdom": "^26.0.0",
-        "nock": "^13.5.5",
-        "prettier": "^3.5.2",
-        "prettier-plugin-organize-imports": "^4.1.0",
-        "react-test-renderer": "^18.3.1",
-        "storybook": "^8.6.3",
-        "tailwindcss": "^4.0.8",
-        "typescript": "^5.6.2",
-        "vite": "^6.0.0",
-        "vite-plugin-html": "^3.2.2",
-        "vitest": "^3.0.6"
-    },
-    "scripts": {
-        "build": "vite build",
-        "build-storybook": "storybook build",
-        "develop": "vite serve",
-        "eslint": "eslint src",
-        "server": "npx tsx  src/server/run.ts",
-        "storybook": "storybook dev -p 6006",
-        "test": "TZ=UTC vitest",
-        "tsc": "tsc --noEmit"
-    },
-    "optionalDependencies": {
-        "fsevents": "^2.3.3"
-    },
-    "prettier": {
-        "tabWidth": 4,
-        "plugins": [
-            "prettier-plugin-organize-imports"
-        ]
-    },
-    "virtool": {
-        "minApiVersion": "23.4.0"
-    }
+  "type": "module",
+  "dependencies": {
+    "@azure/msal-browser": "^2.33.0",
+    "@azure/msal-react": "^1.5.3",
+    "@hookform/resolvers": "^3.9.0",
+    "@radix-ui/react-accordion": "^1.2.0",
+    "@radix-ui/react-checkbox": "^1.1.1",
+    "@radix-ui/react-dialog": "^1.1.1",
+    "@radix-ui/react-dropdown-menu": "^2.1.1",
+    "@radix-ui/react-popover": "^1.1.1",
+    "@radix-ui/react-progress": "^1.1.0",
+    "@radix-ui/react-scroll-area": "^1.1.0",
+    "@radix-ui/react-select": "^2.1.1",
+    "@radix-ui/react-toggle": "^1.1.0",
+    "@radix-ui/react-toggle-group": "^1.1.0",
+    "@radix-ui/react-tooltip": "^1.1.2",
+    "@sentry/react": "^9.0.0",
+    "@tanstack/react-query": "^5.59.0",
+    "class-variance-authority": "^0.7.0",
+    "clsx": "^2.1.1",
+    "commander": "^12.1.0",
+    "d3": "^7.9.0",
+    "date-fns": "^3.6.0",
+    "downshift": "^7.6.2",
+    "express": "^4.21.0",
+    "fuse.js": "^7.1.0",
+    "history": "^4.10.1",
+    "lodash-es": "^4.17.21",
+    "marked": "^14.1.2",
+    "minimist": "^1.2.8",
+    "numbro": "^2.5.0",
+    "radix-ui": "^1.1.3",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-dropzone": "^14.2.3",
+    "react-helmet-async": "^2.0.5",
+    "react-hook-form": "^7.53.0",
+    "react-window": "^1.8.10",
+    "semver": "^7.6.3",
+    "styled-components": "^6.1.13",
+    "superagent": "^8.0.9",
+    "tailwind-merge": "^2.5.2",
+    "tailwindcss-animate": "^1.0.7",
+    "tsx": "^4.19.1",
+    "winston": "^3.17.0",
+    "wouter": "^3.3.5",
+    "yup": "^0.32.11",
+    "zustand": "^4.5.5"
+  },
+  "devDependencies": {
+    "@faker-js/faker": "^9.6.0",
+    "@sentry/vite-plugin": "^3.2.1",
+    "@storybook/addon-actions": "^8.6.12",
+    "@storybook/addon-essentials": "^8.6.12",
+    "@storybook/addon-interactions": "^8.6.12",
+    "@storybook/addon-links": "^8.6.12",
+    "@storybook/preview-api": "^8.6.12",
+    "@storybook/react": "^8.6.12",
+    "@storybook/react-vite": "^8.6.12",
+    "@storybook/test": "^8.6.12",
+    "@tailwindcss/vite": "^4.0.8",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^16.0.1",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/d3": "^7.4.3",
+    "@types/lodash-es": "^4.17.12",
+    "@types/react": "^18.3.5",
+    "@types/react-dom": "^18.3.0",
+    "@types/react-window": "^1.8.8",
+    "@types/styled-components": "^5.1.34",
+    "@types/superagent": "^8.1.9",
+    "@typescript-eslint/eslint-plugin": "^8.5.0",
+    "@typescript-eslint/parser": "^8.5.0",
+    "@vitejs/plugin-react": "^4.3.2",
+    "@vitest/eslint-plugin": "^1.1.32-beta.3",
+    "eslint": "^8.57.1",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "jsdom": "^26.0.0",
+    "nock": "^13.5.5",
+    "prettier": "^3.5.2",
+    "prettier-plugin-organize-imports": "^4.1.0",
+    "react-test-renderer": "^18.3.1",
+    "storybook": "^8.6.12",
+    "tailwindcss": "^4.0.8",
+    "typescript": "^5.6.2",
+    "vite": "^6.0.0",
+    "vite-plugin-html": "^3.2.2",
+    "vitest": "^3.0.6"
+  },
+  "scripts": {
+    "build": "vite build",
+    "build-storybook": "storybook build",
+    "develop": "vite serve",
+    "eslint": "eslint src",
+    "server": "npx tsx  src/server/run.ts",
+    "storybook": "storybook dev -p 6006",
+    "test": "TZ=UTC vitest",
+    "tsc": "tsc --noEmit"
+  },
+  "optionalDependencies": {
+    "fsevents": "^2.3.3"
+  },
+  "prettier": {
+    "tabWidth": 4,
+    "plugins": [
+      "prettier-plugin-organize-imports"
+    ]
+  },
+  "virtool": {
+    "minApiVersion": "23.4.0"
+  }
 }
