+++ src/account/components/AccountGroups.tsx
@@ -4,7 +4,6 @@ import BoxGroupHeader from "@base/BoxGroupHeader";
 import BoxGroupSection from "@base/BoxGroupSection";
 import NoneFound from "@base/NoneFound";
 import { GroupMinimal } from "@groups/types";
-import { map } from "lodash-es";
 import React from "react";
 
 type AccountGroupsProps = {
@@ -21,24 +20,26 @@ export default function AccountGroups({ groups }: AccountGroupsProps) {
             <BoxGroupHeader>
                 <h2>Groups</h2>
             </BoxGroupHeader>
-            <BoxGroupSection className="flex gap-2">
+            <BoxGroupSection className="flex flex-wrap gap-2">
                 {groups.length ? (
-                    map(groups, ({ id, name }) => (
-                        <span
-                            className={cn(
-                                "bg-slate-500",
-                                "flex",
-                                "font-medium",
-                                "items-center",
-                                "text-white",
-                                "px-2 py-1",
-                                "rounded",
-                            )}
-                            key={id}
-                        >
-                            {name}
-                        </span>
-                    ))
+                    groups
+                        .sort((a, b) => a.name.localeCompare(b.name))
+                        .map(({ id, name }) => (
+                            <span
+                                className={cn(
+                                    "bg-slate-500",
+                                    "inline-flex",
+                                    "font-medium",
+                                    "items-center",
+                                    "text-white",
+                                    "px-2 py-1",
+                                    "rounded",
+                                )}
+                                key={id}
+                            >
+                                {name}
+                            </span>
+                        ))
                 ) : (
                     <NoneFound noun="groups" />
                 )}
