+++ src/samples/components/Detail/__tests__/SampleRights.test.tsx
@@ -2,8 +2,9 @@ import { AdministratorRoleName } from "@administration/types";
 import Samples from "@samples/components/Samples";
 import { screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
+import { mockApiListGroups } from "@tests/api/groups";
 import { createFakeAccount, mockApiGetAccount } from "@tests/fake/account";
-import { createFakeGroup, mockApiListGroups } from "@tests/fake/groups";
+import { createFakeGroup } from "@tests/fake/groups";
 import { createFakeSample, mockApiGetSampleDetail } from "@tests/fake/samples";
 import { renderWithRouter } from "@tests/setup";
 import nock from "nock";
