+++ src/tests/fake/groups.ts
@@ -2,7 +2,6 @@ import { faker } from "@faker-js/faker";
 import { Group, GroupMinimal, Permissions } from "@groups/types";
 import { UserNested } from "@users/types";
 import { merge } from "lodash-es";
-import nock from "nock";
 import { createFakePermissions } from "./permissions";
 
 type CreateFakeGroupMinimalProps = {
@@ -48,26 +47,4 @@ export function createFakeGroup(props?: CreateFakeGroupProps): Group {
         permissions: createFakePermissions(permissions),
         users: users || [],
     };
-}
-
-/**
- * Creates a mocked API call for getting a group.
- *
- * @param {Group} group group to be returned from the mocked API call
- * @returns {nock.Scope} nock scope for the mocked API call
- */
-export function mockApiGetGroup(group: Group) {
-    return nock("http://localhost")
-        .get(`/api/groups/${group.id}`)
-        .reply(200, group);
-}
-
-/**
- * Creates a mocked API call for getting a list of groups.
- *
- * @param {Group[]} groups groups to be returned from the mocked API call
- * @returns {nock.Scope} nock scope for the mocked API call
- */
-export function mockApiListGroups(groups: Group[]) {
-    return nock("http://localhost").get("/api/groups").reply(200, groups);
 }
